Move split references to predict
